+++ src/gui-qt4/libs/seiscomp3/gui/core/recordwidget.cpp
@@ -860,6 +860,7 @@ void RecordWidget::init() {
 	_markerSourceWidget = NULL;
 	_filtering = false;
 	_drawMode = Single;
+	_recordBorderDrawMode = SCScheme.records.recordBorders.drawMode;
 	_clipRows = true;
 	_drawOffset = true;
 	_drawRecordID = true;
@@ -899,6 +900,7 @@ void RecordWidget::init() {
 
 	_drawRecords = false;
 	_showAllRecords = false;
+	_showRecordBorders = false;
 	_showScaledValues = false;
 	_autoMaxScale = false;
 	_useGlobalOffset = false;
@@ -3058,7 +3060,9 @@ void RecordWidget::paintEvent(QPaintEvent *event) {
 						}
 						else
 							painter.setPen(offsetColor);
-						painter.drawLine(0,_tracePaintOffset+stream->traces[frontIndex].poly->baseline(), _canvasRect.width(),_tracePaintOffset+stream->traces[frontIndex].poly->baseline());
+
+						painter.drawLine(0,_tracePaintOffset+stream->traces[frontIndex].poly->baseline(),
+						                 _canvasRect.width(),_tracePaintOffset+stream->traces[frontIndex].poly->baseline());
 					}
 
 					if ( stream->antialiasing != isAntialiasing )
@@ -3067,10 +3071,10 @@ void RecordWidget::paintEvent(QPaintEvent *event) {
 					int hMargin = stream->pen.width()-1;
 					if ( hMargin < 0 ) hMargin = 0;
 
-					painter.setPen(_enabled?stream->pen:fg);
-					painter.translate(QPoint(x_tmin[frontIndex], _tracePaintOffset+hMargin));
-					stream->traces[frontIndex].poly->draw(painter);
-					painter.translate(QPoint(-x_tmin[frontIndex], -_tracePaintOffset-hMargin));
+					drawTrace(painter, &stream->traces[frontIndex],
+					          stream->records[frontIndex],
+					          QPen(_enabled?stream->pen:fg),
+					          QPoint(x_tmin[frontIndex], _tracePaintOffset+hMargin));
 				}
 			}
 			break;
@@ -3140,10 +3144,10 @@ void RecordWidget::paintEvent(QPaintEvent *event) {
 					int hMargin = stream->pen.width()-1;
 					if ( hMargin < 0 ) hMargin = 0;
 
-					painter.setPen(_enabled?stream->pen:fg);
-					painter.translate(QPoint(x_tmin[frontIndex], _tracePaintOffset + stream->posY + hMargin));
-					stream->traces[frontIndex].poly->draw(painter);
-					painter.translate(QPoint(-x_tmin[frontIndex], -_tracePaintOffset - stream->posY - hMargin));
+					drawTrace(painter, &stream->traces[frontIndex],
+					          stream->records[frontIndex],
+					          QPen(_enabled?stream->pen:fg),
+					          QPoint(x_tmin[frontIndex], _tracePaintOffset + stream->posY + hMargin));
 				}
 			}
 			break;
@@ -3234,10 +3238,10 @@ void RecordWidget::paintEvent(QPaintEvent *event) {
 					int hMargin = stream->pen.width()-1;
 					if ( hMargin < 0 ) hMargin = 0;
 
-					painter.setPen(_enabled?stream->pen:fg);
-					painter.translate(QPoint(x_tmin[frontIndex], _tracePaintOffset + hMargin));
-					stream->traces[frontIndex].poly->draw(painter);
-					painter.translate(QPoint(-x_tmin[frontIndex], -_tracePaintOffset - hMargin));
+					drawTrace(painter, &stream->traces[frontIndex],
+					          stream->records[frontIndex],
+					          QPen(_enabled?stream->pen:fg),
+					          QPoint(x_tmin[frontIndex], _tracePaintOffset + hMargin));
 				}
 			}
 			break;
@@ -4863,6 +4867,118 @@ void RecordWidget::setData(const QVariant& data) {
 // >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 QVariant RecordWidget::data() const {
 	return _data;
+}
+// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+
+
+
+// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+void RecordWidget::drawTrace(QPainter &painter, Trace *trace,
+                             RecordSequence *seq, const QPen &pen,
+                             const QPoint &paintOffset) {
+	painter.setPen(pen);
+	painter.translate(paintOffset);
+	trace->poly->draw(painter);
+
+	drawRecordBorders(painter, seq);
+
+	painter.translate(paintOffset * -1);
+}
+// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+
+
+
+// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+void RecordWidget::drawRecordBorders(QPainter &painter, RecordSequence *seq) {
+	if ( !_showRecordBorders ) return;
+
+	painter.setRenderHint(QPainter::Antialiasing, false);
+
+	int height = fontMetrics().height() / 3;
+	Core::Time start = leftTime(),
+	           end = rightTime();
+
+	for ( RecordSequence::const_iterator it = seq->begin();
+	      it != seq->end(); ++it ) {
+		const Record *rec = it->get();
+
+		// Skip records that are out of time window [start:end]
+		try {
+			if ( rec->endTime() <= start ) continue;
+		}
+		catch ( ... ) { continue; }
+
+		if ( rec->startTime() >= end ) break;
+
+		int nsamp = rec->sampleCount();
+		if ( nsamp == 0 ) continue;
+
+		int xMin = int(-(_tmin + static_cast<double>((_alignment - rec->startTime()))) *_pixelPerSecond);
+		int xMax = int(-(_tmin + static_cast<double>((_alignment - rec->endTime()))) *_pixelPerSecond);
+
+		QPen pen;
+		QBrush brush;
+		if ( rec->authentication() == Record::SIGNATURE_VALIDATED ) {
+			pen = SCScheme.colors.records.borders.signatureValid.pen;
+			brush = SCScheme.colors.records.borders.signatureValid.brush;
+		}
+		else if ( rec->authentication() == Record::SIGNATURE_VALIDATION_FAILED ) {
+			pen = SCScheme.colors.records.borders.signatureInvalid.pen;
+			brush = SCScheme.colors.records.borders.signatureInvalid.brush;
+		}
+		else {
+			pen = SCScheme.colors.records.borders.standard.pen;
+			brush = SCScheme.colors.records.borders.standard.brush;
+		}
+
+		QRect rect;
+		if ( _recordBorderDrawMode == TopLine ) {
+			rect = QRect(xMin, 0 , xMax - xMin, height);
+		}
+		else if ( _recordBorderDrawMode == BottomLine ) {
+			rect = QRect(xMin, this->height() - height - 1, xMax - xMin, height);
+		}
+		else {
+			rect = QRect(xMin, 0 , xMax - xMin, this->height() - 1);
+		}
+
+		painter.setPen(pen);
+		painter.setBrush(brush);
+		painter.drawRect(rect);
+	}
+}
+// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+
+
+
+// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+void RecordWidget::showRecordBorders(bool enable) {
+	if ( _showRecordBorders == enable ) return;
+
+	_showRecordBorders = enable;
+
+	setDirty();
+	update();
+}
+void RecordWidget::setRecordBorderDrawMode(RecordBorderDrawMode mode) {
+	if ( _recordBorderDrawMode == mode ) return;
+
+	_recordBorderDrawMode = mode;
+
+	setDirty();
+	update();
+}
+// <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+
+
+
+// >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+RecordWidget::RecordBorderDrawMode RecordWidget::recordBorderDrawMode() const {
+	return _recordBorderDrawMode;
 }
 // <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<