+++ src/gui-qt4/libs/seiscomp3/gui/core/recordwidget.h
@@ -219,6 +219,19 @@ class SC_GUI_API RecordWidget : public QWidget {
 			Right
 		};
 
+		enum RecordBorderDrawMode {
+			TopLine,
+			BottomLine,
+			Box
+		};
+
+		struct BorderRect : QRect {
+			BorderRect() : QRect() {}
+			BorderRect(int x, int y, int width, int height) :
+			    QRect(x, y, width, height) {}
+				QColor color;
+		};
+
 		typedef Math::Filtering::InPlaceFilter<float> Filter;
 
 		struct Trace {
@@ -315,6 +328,9 @@ class SC_GUI_API RecordWidget : public QWidget {
 		void setDrawMode(DrawMode mode);
 		DrawMode drawMode() const;
 
+		void setRecordBorderDrawMode(RecordBorderDrawMode mode);
+		RecordBorderDrawMode recordBorderDrawMode() const;
+
 		void setDrawOffset(bool f);
 		bool drawOffset() const { return _drawOffset; }
 
@@ -404,6 +420,9 @@ class SC_GUI_API RecordWidget : public QWidget {
 		//! Whether to show the current selected recordstream or
 		//! both recordstreams
 		void showAllRecords(bool enable);
+
+		//! Whether to show record borders
+		void showRecordBorders(bool enable);
 
 		//! Whether to show scaled or raw values. The default is false.
 		void showScaledValues(bool enable);
@@ -711,6 +730,11 @@ class SC_GUI_API RecordWidget : public QWidget {
 
 		void prepareRecords(Stream *s);
 		void drawRecords(Stream *s, int slot);
+		void drawTrace(QPainter &painter, Trace *trace,
+		               RecordSequence *seq,
+		               const QPen &pen,
+		               const QPoint &paintOffset);
+		void drawRecordBorders(QPainter &painter, RecordSequence *seq);
 
 		void drawAxis(QPainter &painter, const QPen &p);
 
@@ -723,6 +747,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 
 		QVariant             _data;
 		DrawMode             _drawMode;
+		RecordBorderDrawMode _recordBorderDrawMode;
 		Seiscomp::Core::Time _alignment;
 		bool                 _clipRows;
 	
@@ -751,6 +776,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		bool      _drawRecordID;
 		bool      _drawOffset;
 		bool      _showAllRecords;
+		bool      _showRecordBorders;
 		bool      _autoMaxScale;
 		bool      _enabled;
 		bool      _useGlobalOffset;